refector code auth (window)
