+++ lib/GlobalWidget/custom_button.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import '../utils/constant.dart';
+
+class CustomButton extends StatelessWidget {
+  final String routeName;
+  final String text;
+  const CustomButton({Key? key, required this.routeName, required this.text}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    var size = MediaQuery.of(context).size;
+    ButtonStyle buttonStyle = ElevatedButton.styleFrom(
+      onPrimary: Colors.white,
+      primary: kDefualtColorMain,
+      minimumSize: Size(size.width, 50),
+      padding: EdgeInsets.symmetric(horizontal: 16),
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(12)),
+      ),
+    );
+    return ElevatedButton(
+      style: buttonStyle,
+      onPressed: () {
+        Get.toNamed('/signup');
+      },
+      child: Text(
+        "Sign Up",
+        style: TextStyle(
+          fontSize: 18,
+        ),
+      ),
+    );
+  }
+}
