+++ lib/main.dart
@@ -1,26 +1,17 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:petcare_project/screens/before_page.dart';
-import 'package:petcare_project/screens/content_page.dart';
-import 'package:petcare_project/screens/mypet_page.dart';
-import 'package:petcare_project/screens/profile_page.dart';
-import 'package:petcare_project/screens/record_page.dart';
-import 'package:petcare_project/utils/constant.dart';
+import 'package:petcare_project/screens/Auth/before_page.dart';
 import 'package:petcare_project/utils/routes.dart';
-import 'package:petcare_project/utils/bottomnav.dart';
-import 'package:petcare_project/screens/signin_page.dart';
-import 'package:petcare_project/screens/signup_page.dart';
 import 'package:splash_screen_view/SplashScreenView.dart';
 
-void main() {
-  runApp(const MyApp());
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
@@ -37,17 +28,7 @@ class MyApp extends StatelessWidget {
         imageSrc: "assets/image/logo-petcare.png",
         backgroundColor: Colors.white,
       ),
-      // initialRoute: Routes.initialPage,
       getPages: Routes.allRoutes,
-      // routes: {
-      //   '/signin': (context) => SignInPage(),
-      //   '/signup': (context) => SignUpPage(),
-      //   '/route': (context) => RoutePage(),
-      //   '/content': (context) => ContentPage(),
-      //   '/record': (context) => RecordPage(),
-      //   '/mypet': (context) => MyPetPage(),
-      //   '/profile': (context) => ProfilePage(),
-      // },
     );
   }
 }
