+++ lib/screens/Auth/Widget/auth_textfield.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class AuthTextField extends StatelessWidget {
+  final String hintText;
+  final bool obscureText;
+  final TextEditingController controller;
+
+  const AuthTextField({
+    this.obscureText = false,
+    required this.hintText,
+    required this.controller,
+  });
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: Colors.grey[200],
+        border: Border.all(color: Colors.white),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: TextFormField(
+        controller: controller,
+        obscureText: obscureText,
+        decoration: InputDecoration(
+          filled: true,
+          fillColor: Colors.grey.shade200,
+          border: OutlineInputBorder(
+            borderSide: BorderSide.none,
+            borderRadius: BorderRadius.circular(12),
+          ),
+          hintText: hintText,
+        ),
+      ),
+    );
+  }
+}
