+++ lib/screens/Content/content_page.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:google_nav_bar/google_nav_bar.dart';
-import 'package:petcare_project/screens/profile_page.dart';
+import 'package:petcare_project/screens/Profile/profile_page.dart';
 import 'package:petcare_project/utils/constant.dart';
 import 'package:petcare_project/data/menuData.dart';
 import 'package:petcare_project/data/recommendData.dart';
-import 'package:petcare_project/screens/maps_page.dart';
-import 'package:petcare_project/screens/record_page.dart';
+import 'package:petcare_project/screens/Content/maps_page.dart';
+import 'package:petcare_project/screens/Record/record_page.dart';
 
-import '../widget/search_bar.dart';
+import '../../widget/search_bar.dart';
 
 class ContentPage extends StatefulWidget {
   const ContentPage({Key? key}) : super(key: key);
