+++ lib/screens/Content/detaillocation_page.dart
