+++ lib/screens/Profile/profile_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:petcare_project/data/petData.dart';
-import '../utils/constant.dart';
+import '../../utils/constant.dart';
 
 class ProfilePage extends StatefulWidget {
   const ProfilePage({Key? key}) : super(key: key);
