--- lib/screens/otp_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class OtpPage extends StatelessWidget {
-  const OtpPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
-  }
-}
