+++ pubspec.yaml
@@ -42,6 +42,8 @@ dependencies:
   geolocator: ^7.0.2
   readmore: ^2.1.0
   get: ^4.6.5
+  firebase_auth: ^3.4.1
+  firebase_core: ^1.19.1
 
 dev_dependencies:
   flutter_test:
